# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "TJ Online 2nd-hand Trading",
    "description" : "Springboot API Design for Tongji Online Second-hand Trading Platform",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/JEEgroup/JEEwork/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "tags" : [ {
    "name" : "user",
    "description" : "user information"
  }, {
    "name" : "chat",
    "description" : "chatting"
  }, {
    "name" : "favorite",
    "description" : "user's favorite"
  }, {
    "name" : "search",
    "description" : "search for goods"
  } ],
  "paths" : {
    "/chat/addMessage" : {
      "post" : {
        "tags" : [ "chat" ],
        "summary" : "add a new chatting message",
        "operationId" : "addMessage",
        "requestBody" : {
          "description" : "Created a new chatting message",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Chat"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/chat/getContacts/{userId}" : {
      "get" : {
        "tags" : [ "chat" ],
        "summary" : "get all contacts of a user",
        "operationId" : "getContactsById",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "user id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/requestBodies/Contacts"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid userid supplied"
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    },
    "/chat/getMessages/{senderId}/{receiverId}" : {
      "get" : {
        "tags" : [ "chat" ],
        "summary" : "get all messages between two users",
        "operationId" : "getMessages",
        "parameters" : [ {
          "name" : "senderId",
          "in" : "path",
          "description" : "sender user id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "receiverId",
          "in" : "path",
          "description" : "receiver user id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/requestBodies/Contacts"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid userid supplied"
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Chat" : {
        "type" : "object",
        "properties" : {
          "sender_id" : {
            "type" : "integer",
            "example" : 10
          },
          "receiver_id" : {
            "type" : "integer",
            "example" : 13
          },
          "time" : {
            "type" : "string",
            "format" : "date-time"
          },
          "content" : {
            "type" : "string",
            "example" : "hi！"
          }
        },
        "xml" : {
          "name" : "Chat"
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "description" : "User"
      }
    },
    "requestBodies" : {
      "Contacts" : {
        "description" : "All contacts of a user",
        "content" : {
          "application/json" : {
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "required" : true
      },
      "Messages2Users" : {
        "description" : "All messages between two users",
        "content" : {
          "application/json" : {
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Chat"
              }
            }
          }
        },
        "required" : true
      }
    }
  }
}